miningnotify connection hns support
https://github.com/HandshakeAlliance/hstratum/blob/master/docs/spec.md#miningnotify
due to description of hns json parameters changed for the method:miningnotify
When pool sends out the job (mining.notify), it should set following parameters:
Job ID - Used so that stratum can match shares with the client that mined them.
Hash of previous block - Needed in the header.
Coinb1 - The left half of the coinbase transaction. The miner inserts ExtraNonce1 and ExtraNonce2 after this section.
Coinb2 - The right half of the coinbase transaction. This is appended after the miner has added the extra nonce values.
Merkle Tree - This is a list of merkle branches that are hashed along with the newly formed coinbase transaction to get the merkle root.
Tree Root - The root of the Urkel tree that maintains name states. Needed for the block header.
Filter Root - The root of Neutrino filters. Needed for block header.
Reserved Root - A root reserved for future use. Needed for block header.
Block Version - Needed for the block header.
nBits - Needed for the block header. This is the current network difficulty.
nTime - Needed for block header.
clean jobs - If true, informs the miner to release their current work and start on the new job.

works only for f2pool,poolin